Added order field to student stories
Added order field to student stories, didn't need to edit any of the app
code, the server response is just re-ordered.  Kind of an annoying
workaround but strapi doesn't allow re-ordering so we need to do this.
One quirk is that null comes before all numbers, not sure if this is
the desired behaviour or not but it just means that if you need to
make a particular student story first you need to set the order of all
stories.  Not a big deal really just important to note.